FlashCard and FlashCardSet separation
