+++ studyzen/Courses/Requests/CreateFlashCardSetRequest.cs
@@ -1,6 +1,9 @@
 using StudyZen.FlashCards;
 using StudyZen.Common;
 using StudyZen.FlashCards.Requests;
+using StudyZen.FlashCardSets;
+
+
 
 public sealed class CreateFlashCardSetRequest
 {
